requirements.txt
Pillow>=10.0.0
pyobjc>=9.0
pyautogui>=0.9.54
PyYAML>=6.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0
